requirements.txt
alembic==1.17.2
black==25.11.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
filelock==3.20.0
flake8==7.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
future==1.0.0
gevent==25.9.1
greenlet==3.2.4
idna==3.11
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
PySocks==1.7.1
python-dotenv==1.2.1
pytokens==0.3.0
requests==2.32.5
requests-file==3.0.1
SQLAlchemy==2.0.44
SQLAlchemy-Utils==0.42.0
tldextract==5.3.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
zope.event==6.1
zope.interface==8.1
